% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{collection_ids}
\alias{collection_ids}
\title{Collections}
\usage{
collection_ids()
}
\value{
A data frame of available collections, including unique identifier
(\code{collection_id}), title and description.
}
\description{
List of available collections from the \href{https://api.weather.gc.ca/}{MSC GeoMet OGC API} \verb{/collections} endpoint.
}
\examples{
collection_ids()

}
