% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomet-ogc-api.R
\name{geomet_collections}
\alias{geomet_collections}
\title{GeoMet OGC-API}
\usage{
geomet_collections(flatten = FALSE)
}
\arguments{
\item{flatten}{A logical scalar. Should JSON arrays containing only
primitives (i.e. booleans, numbers, and strings) be caused to atomic
vectors? Defaults to \code{FALSE}.}
}
\value{
NULL, an atomic vector, or list.
}
\description{
The \href{https://github.com/opengeospatial/ogcapi-features}{OGC API - Features}
(WFS 3) provides access to collections of geospatial data.
}
\details{
URLs are of the form \verb{/\{collections\}/\{collecion_id\}/items/\{item_id\}?\{query\}}
}
\examples{
\dontrun{
geomet_collections()
}
}
