% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomet-ogc-api.R
\name{geomet_ogc_api}
\alias{geomet_ogc_api}
\title{MSC GeoMet OGC-API Core Request Function}
\usage{
geomet_ogc_api(
  collection_id = NULL,
  items = NULL,
  feature_id = NULL,
  format = "json",
  bbox = NULL,
  properties = NULL,
  ...,
  url_only = FALSE,
  flatten = FALSE
)
}
\arguments{
\item{collection_id}{A character string. Appends URL path. Unique feature collection identifier. Run \code{\link[=collections]{collections()}} to get a list of valid options. If \code{NULL} (default), a list of all available collections from the \href{https://api.weather.gc.ca/}{MSC GeoMet OGC API} \verb{/collections} endpoint.}

\item{items}{A character string. Appends URL path. Set to \code{"items"} to return feature collection features. Defaults to \code{NULL}.}

\item{feature_id}{A character string. Appends URL path. Unique local identifier of the feature. Defaults to \code{NULL}.}

\item{format}{A character string. Appends URL query. API response format. One of \code{"json"}, \code{"jsonld"} or \code{"html"}. Defaults to \code{"json"}.}

\item{bbox}{A character string. Appends URL query. Spatial extent to query. Defaults to \code{NULL}.}

\item{properties}{A character string. Appends URL query. Defaults to \code{NULL}.}

\item{...}{Name-value pairs that provide query parameters. Appends URL query. See \code{\link[httr2:req_url]{httr2::req_url_query()}}.}

\item{url_only}{A logical scalar. If \code{TRUE}, request URL is returned without making GET request to API. Defaults to \code{FALSE}.}

\item{flatten}{A logical scalar. Should JSON arrays containing only primitives (i.e. booleans, numbers, and strings) be caused to atomic vectors? Defaults to \code{FALSE}.}
}
\value{
NULL, an atomic vector, or list.
}
\description{
The \href{https://github.com/opengeospatial/ogcapi-features}{OGC API - Features} (WFS 3) provides access to collections of geospatial data.
URLs are of the form: \verb{https://api.weather.gc.ca/collections/\{collecion_id\}/items/\{item_id\}?\{query\}}.
}
\examples{
\dontrun{
geomet_ogc_api()
}
}
