% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscopendata-package.R
\docType{package}
\name{mscopendata-package}
\alias{mscopendata}
\alias{mscopendata-package}
\title{mscopendata: Functions for interacting and accessing MSC and ECCC Open Data}
\description{
Get present and future weather, water, climate and environmental conditions over Canada and the world.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/miken97/mscopendata}
  \item Report bugs at \url{https://github.com/miken97/mscopendata/issues}
}

}
\author{
\strong{Maintainer}: Mike Nanos \email{mgnanos@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
