% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomet-ogc-api.R
\name{parse_bbox}
\alias{parse_bbox}
\title{Parse bounding box}
\usage{
parse_bbox(coord_vec)
}
\arguments{
\item{coord_vec}{A numeric vector of coordinates with format \code{c(xmin, ymin, xmax, ymax)}.}
}
\value{
A bounding box character string for use in api query.
}
\description{
Parse bounding box
}
\examples{
\dontrun{
bbox <- c("xmin" = -79.44, "ymin" = 43.62, "xmax" = -79.33, "ymax" = 43.68)
parse_bbox(bbox)
}
}
\keyword{internal}
