% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{collections}
\alias{collections}
\title{Available Feature Collections (GeoMet OGC API)}
\usage{
collections(incl_extents = FALSE, url_only = FALSE)
}
\arguments{
\item{incl_extents}{A logical scalar. If \code{TRUE}, spatial and temporal extents for each feature collection are returned. Defaults to \code{FALSE}.}

\item{url_only}{A logical scalar. If \code{TRUE}, request URL is returned without making GET request to API. Defaults to \code{FALSE}.}
}
\value{
A data frame of available features collections, including unique identifier (\code{collection_id}), title and description.
}
\description{
List of available feature collections from the \href{https://api.weather.gc.ca/}{MSC GeoMet OGC API} \verb{/collections} endpoint.
}
\examples{
collections()

}
\seealso{
Other ogc_api: 
\code{\link{geomet_ogc_api}()},
\code{\link{get_collection}()}
}
\concept{ogc_api}
