% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{get_collection}
\alias{get_collection}
\title{Get Feature Collection (GeoMet OGC API)}
\usage{
get_collection(collection_id, use_tidy = TRUE)
}
\arguments{
\item{collection_id}{A character string. Unique collection identifier. Run \code{\link[=collections]{collections()}} to get a list of valid options.}

\item{use_tidy}{A logical scalar. Should API response be tidied? Defaults to \code{TRUE}.}
}
\value{
A data frame of available collections, including unique identifier (\code{collection_id}), title and description.
}
\description{
Get collection from the \href{https://api.weather.gc.ca/}{MSC GeoMet OGC API} \verb{/collections/\{collection_id\}} endpoint.
}
\examples{
\dontrun{
get_collection("climate-stations")
}
}
\seealso{
Other ogc_api: 
\code{\link{collections}()},
\code{\link{geomet_ogc_api}()}
}
\concept{ogc_api}
